feat: [~] move round function by spark needs
