+++ core/src/main/spark_3.0_3.1/scala/doric/syntax/NumericColumns2_31.scala
@@ -4,7 +4,9 @@ package syntax
 import cats.implicits._
 import org.apache.spark.sql.Column
 import org.apache.spark.sql.{functions => f}
-import org.apache.spark.sql.catalyst.expressions.{ShiftLeft, ShiftRight, ShiftRightUnsigned}
+import org.apache.spark.sql.catalyst.expressions.{Expression, RoundBase, ShiftLeft, ShiftRight, ShiftRightUnsigned}
+
+import scala.math.BigDecimal.RoundingMode.RoundingMode
 
 protected trait NumericColumns2_31 {
 
@@ -55,6 +57,33 @@ protected trait NumericColumns2_31 {
       * @see [[org.apache.spark.sql.functions.bitwiseNOT]]
       */
     def bitwiseNot: DoricColumn[T] = column.elem.map(f.bitwiseNOT).toDC
+  }
+
+  /**
+    * NUM WITH DECIMALS OPERATIONS
+    */
+  implicit class NumWithDecimalsOperationsSyntax2_31[T: NumWithDecimalsType](
+      column: DoricColumn[T]
+  ) {
+
+    /**
+      * DORIC EXCLUSIVE! Round the value to `scale` decimal places with given round `mode`
+      * if `scale` is greater than or equal to 0 or at integral part when `scale` is less than 0.
+      *
+      * @todo decimal type
+      * @group Numeric Type
+      */
+    def round(scale: IntegerColumn, mode: RoundingMode): DoricColumn[T] = {
+      case class DoricRound(
+          child: Expression,
+          scale: Expression,
+          mode: RoundingMode
+      ) extends RoundBase(child, scale, mode, s"ROUND_$mode")
+
+      (column.elem, scale.elem)
+        .mapN((c, s) => new Column(DoricRound(c.expr, s.expr, mode)))
+        .toDC
+    }
   }
 
 }
