+++ core/src/main/spark_3.2_3.3_3.4/scala/doric/syntax/NumericColumns32.scala
@@ -4,7 +4,9 @@ package syntax
 import cats.implicits._
 
 import org.apache.spark.sql.{Column, functions => f}
-import org.apache.spark.sql.catalyst.expressions.{ShiftLeft, ShiftRight, ShiftRightUnsigned}
+import org.apache.spark.sql.catalyst.expressions.{Expression, RoundBase, ShiftLeft, ShiftRight, ShiftRightUnsigned}
+
+import scala.math.BigDecimal.RoundingMode.RoundingMode
 
 protected trait NumericColumns32 {
 
@@ -55,6 +57,39 @@ protected trait NumericColumns32 {
       * @see [[org.apache.spark.sql.functions.bitwise_not]]
       */
     def bitwiseNot: DoricColumn[T] = column.elem.map(f.bitwise_not).toDC
+  }
+
+  /**
+    * NUM WITH DECIMALS OPERATIONS
+    */
+  implicit class NumWithDecimalsOperationsSyntax32[T: NumWithDecimalsType](
+      column: DoricColumn[T]
+  ) {
+
+    /**
+      * DORIC EXCLUSIVE! Round the value to `scale` decimal places with given round `mode`
+      * if `scale` is greater than or equal to 0 or at integral part when `scale` is less than 0.
+      *
+      * @todo decimal type
+      * @group Numeric Type
+      */
+    def round(scale: IntegerColumn, mode: RoundingMode): DoricColumn[T] = {
+      case class DoricRound(
+          child: Expression,
+          scale: Expression,
+          mode: RoundingMode
+      ) extends RoundBase(child, scale, mode, s"ROUND_$mode") {
+        override protected def withNewChildrenInternal(
+            newLeft: Expression,
+            newRight: Expression
+        ): DoricRound =
+          copy(child = newLeft, scale = newRight)
+      }
+
+      (column.elem, scale.elem)
+        .mapN((c, s) => new Column(DoricRound(c.expr, s.expr, mode)))
+        .toDC
+    }
   }
 
 }
